Split test into three files
That way we escape the problem that implicit timeouts are reduced too much
for later tests.
