+++ tests/neg/missing-implicit-2.check
@@ -1,0 +1,28 @@
+-- Error: tests/neg/missing-implicit-2.scala:4:24 ----------------------------------------------------------------------
+4 |val f = Future[Unit] { } // error
+  |                        ^
+  |                     Cannot find an implicit ExecutionContext. You might pass
+  |                     an (implicit ec: ExecutionContext) parameter to your method.
+  |
+  |                     The ExecutionContext is used to configure how and on which
+  |                     thread pools Futures will run, so the specific ExecutionContext
+  |                     that is selected is important.
+  |
+  |                     If your application does not define an ExecutionContext elsewhere,
+  |                     consider using Scala's global ExecutionContext by defining
+  |                     the following:
+  |
+  |                     implicit val ec: scala.concurrent.ExecutionContext = scala.concurrent.ExecutionContext.global
+  |
+  |                     The following import might fix the problem:
+  |
+  |                       import concurrent.ExecutionContext.Implicits.global
+  |
+-- [E007] Type Mismatch Error: tests/neg/missing-implicit-2.scala:6:25 -------------------------------------------------
+6 |val b: java.lang.Byte = (1: Byte) // error, but no hint
+  |                         ^^^^^^^
+  |                         Found:    Byte
+  |                         Required: Byte²
+  |
+  |                         where:    Byte  is a class in package scala
+  |                                   Byte² is a class in package java.lang
