+++ tests/neg/missing-implicit-3.check
@@ -1,0 +1,25 @@
+-- [E007] Type Mismatch Error: tests/neg/missing-implicit-3.scala:6:44 -------------------------------------------------
+6 |val d: scala.concurrent.duration.Duration = (10, DAYS) // error
+  |                                            ^^^^^^^^^^
+  |                                            Found:    (Int, java.util.concurrent.TimeUnit)
+  |                                            Required: concurrent².duration.Duration
+  |
+  |                                            where:    concurrent  is a package in package java.util
+  |                                                      concurrent² is a package in package scala
+  |
+  |
+  |                                            The following import might fix the problem:
+  |
+  |                                              import concurrent.duration.pairIntToDuration
+  |
+-- [E008] Not Found Error: tests/neg/missing-implicit-3.scala:8:48 -----------------------------------------------------
+8 |val d2: scala.concurrent.duration.Duration = 10.days // error
+  |                                             ^^^^^^^
+  |                            value days is not a member of Int, but could be made available as an extension method.
+  |
+  |                            One of the following imports might fix the problem:
+  |
+  |                              import concurrent.duration.DurationInt
+  |                              import concurrent.duration.DurationLong
+  |                              import concurrent.duration.DurationDouble
+  |
